Move database to postgres

+++ database/migrations/20240323211220-create-card-table.js
@@ -2394,15 +2394,15 @@ module.exports = {
         allowNull: true,
       },
       cost: {
-        type: Sequelize.TINYINT,
+        type: Sequelize.INTEGER,
         allowNull: true,
       },
       attack: {
-        type: Sequelize.TINYINT,
+        type: Sequelize.INTEGER,
         allowNull: true,
       },
       health: {
-        type: Sequelize.TINYINT,
+        type: Sequelize.INTEGER,
         allowNull: true,
       },
       image: Sequelize.STRING,
